ftests: ftest-wrapper - add trap function for GitHub workflow
When the GitHub workflow is canceled, the VMs will have taken
github-runner.lock but would not release it because of the cleanup()
the function doesn't get called on such events, starving the next CI
jobs, which waits to acquire the lock before starting its run.
Fix it by adding a trap function for SIGINT and SIGTERM, that calls
the cleanup(), which releases the github-runner.lock.

As per
https://docs.github.com/en/actions/managing-workflow-runs/canceling-a-workflow
SIGINT and SIGTERM are the signals used by GitHub to cancel the bash
scripts used in the workflow steps, hence install trap for those
signals.

Signed-off-by: Kamalesh Babulal <kamalesh.babulal@oracle.com>