cgrulesengd: Fix the return value check in recvfrom
In cgre_receive_netlink_msg::recvfrom(), the return value recv_len is
checked against ENOBUFS, which is an errno value that represents the
last occurred error. However, this comparison has two issues:

First, if the recv_len message bytes are equal to the value of ENOBUFS,
it will return a false positive error.

Second, recv_len is not checked for -1, which indicates an unexpected
event that requires investigation. Fortunately, the next line acts as a
safety net, returning on error conditions and EOF.

Fix this by checking for -1 (error) and comparing errno against ENOBUFS.

Signed-off-by: Kamalesh Babulal <kamalesh.babulal@oracle.com>